requirements.txt
beautifulsoup4>=4.12.0
requests>=2.31.0
playwright>=1.40.0
selenium>=4.15.0
redis>=5.0.0
pytest>=7.4.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
lxml>=4.9.0
html2text>=2020.1.16
jsonschema>=4.17.0
